requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
boto3==1.42.63
botocore==1.42.63
cachetools==7.0.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
cvdata==0.0.3
flatbuffers==25.12.19
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.14.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
libclang==18.1.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.17.0
numpy==2.4.2
opencv-python==4.13.0.92
openimages==0.0.1
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
s3transfer==0.16.0
setuptools==82.0.0
six==1.17.0
smmap==5.0.2
streamlit==1.55.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
wrapt==2.1.2
